requirements.txt
streamlit
opencv-python
numpy
Pillow
joblib